Rename components to be more descriptive.
